some fixes and changes
